# CUDA: faster q8_0 -> f16 dequantization