This PR adds a faster kernel for q8_0 -> FP16 conversion. It works by first copying the quantized data as ints to shared memory to ensure that the global memory reads are fully coalesced. It then reads the data from shared memory, dequantizes it, and writes the result back as half2 (again coalesced). To make this work a single warp works on 2048 data values each; with this many values there are 2048/32 = 64 q8_0 blocks with a size of 34\*64 = 2176 bytes which is divisible by 32\*sizeof(int) = 128 bytes and thus allows for aligned reads.

The new kernel is 1.64 times faster. For perplexity calculations with a context size of 4096 the overall performance changes as follows:

| GPU      | Model         | Batch size | t/s master | t/s PR | Speedup |
|----------|---------------|------------|------------|--------|---------|
| RTX 3090 | llama 7B Q8_0 |         64 |       1073 |   1103 |    1.03 |
| RTX 3090 | llama 7B Q8_0 |        128 |       1724 |   1788 |    1.04 |
| RTX 3090 | llama 7B Q8_0 |        256 |       2357 |   2472 |    1.05 |
| RTX 3090 | llama 7B Q8_0 |        512 |       2716 |   2848 |    1.05 |
| RTX 3090 | llama 7B Q8_0 |       1024 |       2981 |   3146 |    1.06 |
| RTX 3090 | llama 7B Q8_0 |       2048 |       2979 |   3084 |    1.04 |
| RTX 3090 | llama 7B Q8_0 |       4096 |       2783 |   2837 |    1.02 |

The improvements are largest for medium batch sizes because for large batch sizes the dequantization is negligible while for small batch sizes the FP16 matrix multiplication is comparatively slow.

I used perplexity calculations for benchmarking because `llama-bench` gave me suspiciously small values for how much the speed improves:

| GPU      | Model         |   Batch size | Test   |   t/s master |   t/s cuda-faster-q8_0-to-half-2 |   Speedup |
|:---------|:--------------|-------------:|:-------|-------------:|---------------------------------:|----------:|
| RTX 3090 | llama 7B Q8_0 |           64 | pp4096 |      1098.45 |                          1121.73 |      1.02 |
| RTX 3090 | llama 7B Q8_0 |          128 | pp4096 |      1783.72 |                          1817.29 |      1.02 |
| RTX 3090 | llama 7B Q8_0 |          256 | pp4096 |      2492.58 |                          2534.03 |      1.02 |
| RTX 3090 | llama 7B Q8_0 |          512 | pp4096 |      2897.28 |                          2917.15 |      1.01 |
| RTX 3090 | llama 7B Q8_0 |         1024 | pp4096 |      3212.01 |                          3229.19 |      1.01 |
| RTX 3090 | llama 7B Q8_0 |         2048 | pp4096 |      3162.22 |                          3172.89 |      1.00 |
| RTX 3090 | llama 7B Q8_0 |         4096 | pp4096 |      2909.47 |                          2912.44 |      1.00 |

Maybe there's a bug somewhere in `llama-bench` that causes the results to not be 100% representative?