@kalomaze 2024-01-12
> Maybe there's a bug somewhere in `llama-bench` that causes the results to not be 100% representative?

It evaluates the same token repeatedly, which also makes it inaccurate for MoE analysis due to unrealistic memory access patterns