@slaren 2024-01-12
> Maybe there's a bug somewhere in `llama-bench` that causes the results to not be 100% representative?

I can think of two differences:
- `llama-bench` always evaluates BOS tokens, but I don't see how that could make a difference in Q8_0 quantization
- `llama-bench` does 5 repetitions, which may be fewer than the number of chunks you tested with perplexity. The first run is likely to be slower than the rest due cold caches, and increasing the number of repetitions would reduce the effect of the first run.