@slaren 2024-01-12
> is it irrelevant for full offload in this case?

Even with full offload, the `token_embd` tensor is kept on the CPU, and is loaded with mmap if enabled. So it is not impossible that it could affect, but I was thinking about CUDA caches. You can easily test that with `-no-mmap`.