@slaren 2024-01-12
I merged this on top of #4766 and tested the Q8_0 F32 mat mul performance with test-backend-ops, and with llama-bench.

test-backend-ops:

| TEST           | prev GB/s   | PR GB/s   | speedup |
|----------------|-------------|-----------|---------|
| 4096x4096x64   | 41933.55    | 43481.98  | 1.036 |
| 4096x4096x128  | 69505.51    | 71041.41  | 1.022 |
| 4096x4096x256  | 126507.85   | 129666.49 | 1.024 |
| 4096x4096x512  | 155679.53   | 156673.03 | 1.006 |
| 4096x4096x1024 | 201456.24   | 203123.04 | 1.008 |
| 4096x4096x2048 | 224171.62   | 224793.63 | 1.002 |
| 4096x4096x4096 | 246820.14   | 247262.70 | 1.001 |
| 4096x11008x64  | 45590.37    | 47138.21  | 1.033 |
| 4096x11008x128 | 89164.12    | 90921.79  | 1.019 |
| 4096x11008x256 | 144903.57   | 147355.57 | 1.016 |
| 4096x11008x512 | 170333.41   | 174329.86 | 1.023 |
| 4096x11008x1024| 228296.06   | 230166.65 | 1.008 |
| 4096x11008x2048| 252775.76   | 253955.23 | 1.004 |
| 4096x11008x4096| 275726.27   | 276119.70 | 1.001 |



llama-bench:

| TEST          | prev t/s   | PR t/s    | speedup |
|---------------|------------|-----------|---------|
| n_batch 64    | 1164.85    | 1185.82   | 1.018 |
| n_batch 128   | 1935.84    | 1960.78   | 1.012 |
| n_batch 256   | 2825.21    | 2834.42   | 1.003 |
| n_batch 512   | 3336.03    | 3371.92   | 1.010 |
| n_batch 1024  | 3772.85    | 3785.25   | 1.003 |
| n_batch 2048  | 3762.17    | 3775.87   | 1.003 |
| n_batch 4096  | 3426.63    | 3432.82   | 1.001 |



<details>

prev:
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=64,k=4096,bs=[1,1],nr=[1,1]):                20 runs -   120.75 us/run -  5309440 kB/run - 41933.55 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=128,k=4096,bs=[1,1],nr=[1,1]):               20 runs -   145.70 us/run - 10618880 kB/run - 69505.51 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=256,k=4096,bs=[1,1],nr=[1,1]):               20 runs -   160.10 us/run - 21237760 kB/run - 126507.85 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=512,k=4096,bs=[1,1],nr=[1,1]):               20 runs -   260.20 us/run - 42475520 kB/run - 155679.53 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=1024,k=4096,bs=[1,1],nr=[1,1]):              20 runs -   402.15 us/run - 84951040 kB/run - 201456.24 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=2048,k=4096,bs=[1,1],nr=[1,1]):              20 runs -   722.80 us/run - 169902080 kB/run - 224171.62 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=4096,k=4096,bs=[1,1],nr=[1,1]):              20 runs -  1312.95 us/run - 339804160 kB/run - 246820.14 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=64,k=11008,bs=[1,1],nr=[1,1]):               20 runs -   298.45 us/run - 14267392 kB/run - 45590.37 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=128,k=11008,bs=[1,1],nr=[1,1]):              20 runs -   305.20 us/run - 28534784 kB/run - 89164.12 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=256,k=11008,bs=[1,1],nr=[1,1]):              20 runs -   375.60 us/run - 57069568 kB/run - 144903.57 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=512,k=11008,bs=[1,1],nr=[1,1]):              20 runs -   639.05 us/run - 114139136 kB/run - 170333.41 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=1024,k=11008,bs=[1,1],nr=[1,1]):             20 runs -   953.60 us/run - 228278272 kB/run - 228296.06 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=2048,k=11008,bs=[1,1],nr=[1,1]):             20 runs -  1722.50 us/run - 456556544 kB/run - 252775.76 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=4096,k=11008,bs=[1,1],nr=[1,1]):             20 runs -  3158.25 us/run - 913113088 kB/run - 275726.27 GB/s


PR:
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=64,k=4096,bs=[1,1],nr=[1,1]):                20 runs -   116.45 us/run -  5309440 kB/run - 43481.98 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=128,k=4096,bs=[1,1],nr=[1,1]):               20 runs -   142.55 us/run - 10618880 kB/run - 71041.41 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=256,k=4096,bs=[1,1],nr=[1,1]):               20 runs -   156.20 us/run - 21237760 kB/run - 129666.49 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=512,k=4096,bs=[1,1],nr=[1,1]):               20 runs -   258.55 us/run - 42475520 kB/run - 156673.03 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=1024,k=4096,bs=[1,1],nr=[1,1]):              20 runs -   398.85 us/run - 84951040 kB/run - 203123.04 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=2048,k=4096,bs=[1,1],nr=[1,1]):              20 runs -   720.80 us/run - 169902080 kB/run - 224793.63 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=4096,k=4096,bs=[1,1],nr=[1,1]):              20 runs -  1310.60 us/run - 339804160 kB/run - 247262.70 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=64,k=11008,bs=[1,1],nr=[1,1]):               20 runs -   288.65 us/run - 14267392 kB/run - 47138.21 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=128,k=11008,bs=[1,1],nr=[1,1]):              20 runs -   299.30 us/run - 28534784 kB/run - 90921.79 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=256,k=11008,bs=[1,1],nr=[1,1]):              20 runs -   369.35 us/run - 57069568 kB/run - 147355.57 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=512,k=11008,bs=[1,1],nr=[1,1]):              20 runs -   624.40 us/run - 114139136 kB/run - 174329.86 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=1024,k=11008,bs=[1,1],nr=[1,1]):             20 runs -   945.85 us/run - 228278272 kB/run - 230166.65 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=2048,k=11008,bs=[1,1],nr=[1,1]):             20 runs -  1714.50 us/run - 456556544 kB/run - 253955.23 GB/s
MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=4096,k=11008,bs=[1,1],nr=[1,1]):             20 runs -  3153.75 us/run - 913113088 kB/run - 276119.70 GB/s



Device 0: NVIDIA GeForce RTX 3090 Ti, compute capability 8.6, VMM: yes
| model                          |       size |     params | backend    | ngl |    n_batch | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |         64 | pp 4096    |  1164.85 ± 18.29 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |        128 | pp 4096    |  1935.84 ± 11.70 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |        256 | pp 4096    |  2825.21 ± 14.31 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |        512 | pp 4096    |  3336.03 ± 20.44 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |       1024 | pp 4096    |  3772.85 ± 10.57 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |       2048 | pp 4096    |   3762.17 ± 9.95 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |       4096 | pp 4096    |   3426.63 ± 8.73 |

build: 53ae0dd8 (1879)

ggml_init_cublas: found 1 CUDA devices:
  Device 0: NVIDIA GeForce RTX 3090 Ti, compute capability 8.6, VMM: yes
| model                          |       size |     params | backend    | ngl |    n_batch | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |         64 | pp 4096    |  1185.82 ± 11.26 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |        128 | pp 4096    |  1960.78 ± 18.39 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |        256 | pp 4096    |  2834.42 ± 21.17 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |        512 | pp 4096    |  3371.92 ± 14.23 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |       1024 | pp 4096    |  3785.25 ± 13.56 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |       2048 | pp 4096    |   3775.87 ± 6.65 |
| llama 7B Q8_0                  |   6.67 GiB |     6.74 B | CUDA       |  99 |       4096 | pp 4096    |   3432.82 ± 9.70 |

build: 28d1ec4c (1881)


</details>